backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9
pgvector==0.2.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data Validation & Settings
pydantic==2.5.0
pydantic-settings==2.1.0

# AI/ML for Face Recognition - Using InsightFace (better than face_recognition)
opencv-python==4.8.1.78
insightface==0.7.3
onnxruntime==1.16.3  # Required for InsightFace
numpy==1.24.3
pillow==10.1.0
scikit-learn==1.3.2
# face-recognition==1.3.0  # Replaced with InsightFace

# Utilities
python-dotenv==1.0.0
email-validator==2.1.0
aiofiles==23.2.1

# CORS for React integration
fastapi-cors==0.0.6

# PDF Generation
reportlab==4.0.7
weasyprint==60.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
